Python RFQ/requirements.txt
fastapi
uvicorn
jinja2
asyncpg
python-dotenv
python-multipart
pydantic-settings
bcrypt
pyjwt
gunicorn
